# plan: support subquery in `Do` statement