<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Before this PR:
```
mysql> select * from t1;
+------+
| id   |
+------+
|    1 |
+------+
1 row in set (0.01 sec)

mysql> select 1 in (select * from t1);
+-------------------------+
| 1 in (select * from t1) |
+-------------------------+
|                       1 |
+-------------------------+
1 row in set (0.00 sec)

mysql> do 1 in (select * from t1);
ERROR 2013 (HY000): Lost connection to MySQL server during query
```
while in MySQL:
```
mysql> select * from t1;
+--------------+
| processor_id |
+--------------+
|            1 |
+--------------+
1 row in set (0.00 sec)

mysql> do 1 in (select * from t1);
Query OK, 0 rows affected (0.00 sec)
```

MySQL states in https://dev.mysql.com/doc/refman/5.7/en/do.html that:
> DO only executes expressions. It cannot be used in all cases where SELECT can be used. For example, DO id FROM t1 is invalid because it references a table.

The documentation does not explicitly specify the boundary to which `Do` is compatible with `SELECT`. Since this case experimentally works in MySQL, we support it in TiDB.

### What is changed and how it works?

- involve full expression rewriting in plan building for `Do`;
- prevent `Projection` specially added for `Do` being eliminated in optimizer;

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Side effects

N/A

Related changes

 - Need to cherry-pick to the release branch

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/pingcap/tidb/8343)
<!-- Reviewable:end -->