@eurekaka 2018-11-16
/run-all-tests